# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bookstore API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface Book {
  id: number;
  title: string;
  description: string;
  author: string;
  price: number;
  salePrice?: number | null;
  isOnSale: boolean;
  isFree: boolean;
  isFeatured: boolean;
  coverImage: string;
  category: string;
  language: string;
  ageGroup: string;
  pages?: number | null;
  downloadUrl?: string | null;
  stock: number;
  rating: number;
  reviewCount: number;
  createdAt: string;
}

export interface BookListResponse {
  books: Book[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateBookBody {
  title: string;
  description: string;
  author: string;
  price: number;
  salePrice?: number | null;
  isOnSale?: boolean;
  isFree?: boolean;
  isFeatured?: boolean;
  coverImage: string;
  category: string;
  language: string;
  ageGroup: string;
  pages?: number | null;
  downloadUrl?: string | null;
  stock?: number;
}

export interface UpdateBookBody {
  title?: string;
  description?: string;
  author?: string;
  price?: number;
  salePrice?: number | null;
  isOnSale?: boolean;
  isFree?: boolean;
  isFeatured?: boolean;
  coverImage?: string;
  category?: string;
  language?: string;
  ageGroup?: string;
  pages?: number | null;
  downloadUrl?: string | null;
  stock?: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  bookCount: number;
}

export interface OrderItem {
  bookId: number;
  title: string;
  price: number;
  quantity: number;
  coverImage: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  city: string;
  paymentMethod: string;
  paymentReference?: string | null;
  status: string;
  total: number;
  items: OrderItem[];
  createdAt: string;
  notes?: string | null;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateOrderBodyItemsItem = {
  bookId: number;
  quantity: number;
};

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  city: string;
  paymentMethod: string;
  paymentReference?: string | null;
  notes?: string | null;
  items: CreateOrderBodyItemsItem[];
}

export interface UpdateOrderStatusBody {
  status: string;
}

export interface AdminStats {
  totalBooks: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  completedOrders: number;
  freeBooks: number;
  booksOnSale: number;
  recentOrders: Order[];
  topSellingBooks: Book[];
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
  message: string;
}

export type ListBooksParams = {
  category?: string;
  language?: string;
  isFree?: boolean;
  isOnSale?: boolean;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListOrdersParams = {
  status?: string;
  page?: number;
};