# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bookstore API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: books
    description: Book operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin operations
  - name: categories
    description: Category operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: isFree
          in: query
          schema:
            type: boolean
        - name: isOnSale
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"
    post:
      operationId: createBook
      tags: [admin]
      summary: Create a new book (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a single book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    put:
      operationId: updateBook
      tags: [admin]
      summary: Update a book (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [admin]
      summary: Delete a book (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /books/featured:
    get:
      operationId: getFeaturedBooks
      tags: [books]
      summary: Get featured/bestseller books
      responses:
        "200":
          description: Featured books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /books/on-sale:
    get:
      operationId: getOnSaleBooks
      tags: [books]
      summary: Get books currently on sale
      responses:
        "200":
          description: Sale books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /books/free:
    get:
      operationId: getFreeBooks
      tags: [books]
      summary: Get free books/articles
      responses:
        "200":
          description: Free books and articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        author:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        isOnSale:
          type: boolean
        isFree:
          type: boolean
        isFeatured:
          type: boolean
        coverImage:
          type: string
        category:
          type: string
        language:
          type: string
        ageGroup:
          type: string
        pages:
          type: integer
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - author
        - price
        - isOnSale
        - isFree
        - isFeatured
        - coverImage
        - category
        - language
        - ageGroup
        - stock
        - rating
        - reviewCount
        - createdAt

    BookListResponse:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - books
        - total
        - page
        - totalPages

    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        author:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        isOnSale:
          type: boolean
          default: false
        isFree:
          type: boolean
          default: false
        isFeatured:
          type: boolean
          default: false
        coverImage:
          type: string
        category:
          type: string
        language:
          type: string
        ageGroup:
          type: string
        pages:
          type: integer
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        stock:
          type: integer
          default: 100
      required:
        - title
        - description
        - author
        - price
        - coverImage
        - category
        - language
        - ageGroup

    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        author:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        isOnSale:
          type: boolean
        isFree:
          type: boolean
        isFeatured:
          type: boolean
        coverImage:
          type: string
        category:
          type: string
        language:
          type: string
        ageGroup:
          type: string
        pages:
          type: integer
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        stock:
          type: integer

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        bookCount:
          type: integer
      required:
        - id
        - name
        - slug
        - bookCount

    OrderItem:
      type: object
      properties:
        bookId:
          type: integer
        title:
          type: string
        price:
          type: number
        quantity:
          type: integer
        coverImage:
          type: string
      required:
        - bookId
        - title
        - price
        - quantity
        - coverImage

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        city:
          type: string
        paymentMethod:
          type: string
        paymentReference:
          type: string
          nullable: true
        status:
          type: string
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - address
        - city
        - paymentMethod
        - status
        - total
        - items
        - createdAt

    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages

    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        city:
          type: string
        paymentMethod:
          type: string
        paymentReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              bookId:
                type: integer
              quantity:
                type: integer
            required:
              - bookId
              - quantity
      required:
        - customerName
        - customerEmail
        - customerPhone
        - address
        - city
        - paymentMethod
        - items

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AdminStats:
      type: object
      properties:
        totalBooks:
          type: integer
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        completedOrders:
          type: integer
        freeBooks:
          type: integer
        booksOnSale:
          type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        topSellingBooks:
          type: array
          items:
            $ref: "#/components/schemas/Book"
      required:
        - totalBooks
        - totalOrders
        - totalRevenue
        - pendingOrders
        - completedOrders
        - freeBooks
        - booksOnSale
        - recentOrders
        - topSellingBooks

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        message:
          type: string
      required:
        - success
        - token
        - message